incident: Move `choices` out of `models` module
This lets us import the choices without importing the rest of
`incident.models.__init__`, which can cause circular imports during
Django's model initialization.

Technically, choices aren't models.  The less we have in the `models`
module, the better (IMO), and even though choices are _related_ to
models, they're not actual model classes, so let's not overburden
ourselves here.
