+++ incident/tests/test_models.py
@@ -2,7 +2,7 @@ from datetime import date
 
 from django.test import TestCase
 
-from incident.models import choices
+from incident import choices
 from .factories import (
     IncidentChargeFactory,
     IncidentChargeWithUpdatesFactory,
